+++ pkg/app/dispatcher.go
@@ -99,7 +99,7 @@ type ClientDispatcher interface {
 // NewClientTester creates a testing dispatcher that simulates a
 // client environment. The given UI element is mounted upon creation.
 func NewClientTester(n UI) ClientDispatcher {
-	e := &engineX{
+	e := &engine{
 		ActionHandlers: actionHandlers,
 	}
 	e.init()
@@ -129,7 +129,7 @@ type ServerDispatcher interface {
 // NewServerTester creates a testing dispatcher that simulates a
 // client environment.
 func NewServerTester(n UI) ServerDispatcher {
-	e := &engineX{
+	e := &engine{
 		IsServerSide:   true,
 		ActionHandlers: actionHandlers,
 	}
